split build_chunks worker into a different process
